Commits

+++ backend/graphql/resolvers/mutations/productResolvers.js
@@ -1,14 +1,25 @@
 import Product from '../../../database/models/product'
+import User from '../../../database/models/user'
 import { pick } from '../../../utils'
+import {
+  ADD_PRODUCT,
+  REMOVE_PRODUCT,
+  UPDATE_PRODUCT_INFO,
+  UPDATE_PRODUCT_IMAGES,
+  UPDATE_PRODUCT_QUANTITY,
+} from '../../../database/operations'
 
 const addProductResolver = async (parent, args, context) => {
   // TODO: Check for admin authorization here.
   const { user } = context
   if (!user) {
-    throw new Error('Unauthorized')
+    throw new Error('Unauthenticated')
   }
 
   try {
+    if (!user.isAuthorizedTo(ADD_PRODUCT)) {
+      throw new Error('Unauthorized')
+    }
     const product = await new Product(args).save()
     return product
   } catch (err) {
@@ -17,13 +28,15 @@ const addProductResolver = async (parent, args, context) => {
 }
 
 const removeProductResolver = async (parent, args, context) => {
-  // TODO: Check for admin authorization here.
   const { user } = context
   if (!user) {
-    throw new Error('Unauthorized')
+    throw new Error('Unauthenticated')
   }
 
   try {
+    if (!user.isAuthorizedTo(REMOVE_PRODUCT)) {
+      throw new Error('Unauthorized')
+    }
     const removedProduct = await Product.findByIdAndRemove(args.id)
     return removedProduct
   } catch (err) {
@@ -32,13 +45,15 @@ const removeProductResolver = async (parent, args, context) => {
 }
 
 const updateProductInfoResolver = async (parent, args, context) => {
-  // TODO: Check for admin authorization here.
   const { user } = context
   if (!user) {
-    throw new Error('Unauthorized')
+    throw new Error('Unauthenticated')
   }
 
   try {
+    if (!user.isAuthorizedTo(UPDATE_PRODUCT_INFO)) {
+      throw new Error('Unauthorized')
+    }
     const productArgs = pick(args, [
       'name',
       'category',
@@ -63,10 +78,13 @@ const updateProductImagesResolver = async (parent, args, context) => {
   // TODO: Check for admin authorization here.
   const { user } = context
   if (!user) {
-    throw new Error('Unauthorized')
+    throw new Error('Unauthenticated')
   }
 
   try {
+    if (!user.isAuthorizedTo(UPDATE_PRODUCT_IMAGES)) {
+      throw new Error('Unauthorized')
+    }
     const productArgs = pick(args, ['imagePath'])
     const product = await Product.findByIdAndUpdate(args.id, productArgs, {
       new: true,
@@ -84,8 +102,10 @@ const updateProductQuantityResolver = async (parent, args, context) => {
   if (!user) {
     throw new Error('Unauthorized')
   }
-
   try {
+    if (!user.isAuthorizedTo(UPDATE_PRODUCT_QUANTITY)) {
+      throw new Error('Unauthorized')
+    }
     const productArgs = pick(args, ['size'])
 
     if (!productArgs.size) {
